Create README.md
This is a To do list app made in react for managing user's to do list. User can add the task and the date assigned for that task to complete, once the task is finsished the task can be deleted from the to do list.

This project is made in react with react hooks, the intention behind this project was to gain the practical experience while working on the project.